requirements.txt
Django==4.2.6
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
django-cors-headers==4.7.0
django-unfold==0.48.0
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
whitenoise==6.9.0
